Add limited HTTP/2 support, with a notable caveat (#11)
To support bootstrap IP addresses, we moved away from using URLSession
as that does not let you override name resolution of the URL - instead
we managed the connection ourselves and use CFHTTPMessage to build and
parse the HTTP request and response. However, CFHTTPMessage is limited
to HTTP/1.1 and does not actually support HTTP/2.

Supporting HTTP/2 with bootstrap IP addresses is going to be complex and
likely require pulling in an external dependency of some kind. So, until
that is sorted out, I'm exposing a new option in TransportOptions to
specify if HTTP/2 is desired (the default), but noting that if bootstrap
IP address is defined, HTTP/1.1 will be used.